fix(api): several engine and api stacker fixes
Two big things:
- The core actually needs to make new labware cores sometimes, which is
pretty easy to accomplish. We're getting the labware by asking the
engine "hey what's on this module" after the command completes (rather
than by getting the results of the command) because if the command
failed and went through error recovery, its results may not be reliable.
- More seriously, retrieve() needs to simultaneously load multiple
labware. To do this, it uses internal engine methods that inspect what
the position of the loaded labware is and walks the parent list down to
the deck. So if you load labware A on location L and labware B on
labware A simultaneously, how can you do this? The answer is probably
"have a mechanism for doing state updates ahead of time as a temporary
and operate on the projected state", but that's a lot so the answer for
now is "the affected methods get a projection of a slice of state".