feat(protocol-designer): generate Python for pickUpTip command (#17610)
# Overview

This emits Python for the PD atomic pickUpTip command. The Python code
looks like this:
```
pipette_left.pick_up_tip(location=tip_rack_1)
```

PD implements its own tip tracking and calculates exactly which tip to
pick up. However, we don't want to emit the exact tip well in the Python
code because it would make it painful for someone to edit the Python
code to add or remove steps.

We do emit the tiprack ID in the Python code because PD makes the user
choose which tiprack to use for each transfer. If we didn't emit the
tiprack ID, the PAPI would pick the next available tiprack, which may
not match the tiprack that the user chose in PD.

AUTH-1093

## Test Plan and Hands on Testing

There was no existing unit test for the JSON `pickUpTip` command, so I
added one.

With this change, we can now generate a runnable Python protocol for a
simple PD transfer. I checked that the protocol passes with `simulate`.

## Risk assessment

Low. Python generation is not used externally yet.