app: fix load-module AA handling
We do deck configuration conflict checking by accumulating all used
addressable areas; figuring out every set of cutout fixtures that could
provide those addressable areas; and then figuring out if there are any
cutout fixtures loaded currently that are not in any of those sets of
cutout fixtures.

We accumulate addressable areas differently in the engine and the app.
In the engine, as a command decides what to do and which addressable
areas it will touch, it marks those addressable areas as used. It can do
that because, well, it's the thing touching them. And it can and will
mark an addressable area as used if it's used, _even if it's not
apparent to anything outside the engine that this addressable area is in
fact used_.

But in the app, we can't see inside the engine. We can only see the
command parameters and the command results. So we emulate this, but we
can't emulate it fully, because there are some commands that touch
addressable areas that don't show up in their commands or results - for
instance, the absorbance reader open and close lid commands.

So the app, specifically for modules because that's where the problem
is, was essentially doing deck configuration twice for modules: when it
saw a module loaded, it would figure out the addressable area for that
module (which isn't in the command) by loading the cutout fixture for
that module into the cutout that module's location specified and then
marking all the addressable areas that the module fixture provides when
loaded into that addressable area as used. Meantime, the engine is still
just marking the addressable areas that are actually used as used.

This worked until we added these combo fixtures that fuse the stacker
with other fixtures, like the magblock or the waste chute. Now, when the
app proactively loads the two addressable areas provided by the default
stacker fixture, there is no resolvable set of cutout fixtures that can
also provide the addressable areas used by the combo fixtures: nothing
can be provide D3, _and_ flexStackerModuleV1D4, _and_ gripperWasteChute,
because D3 can't be there. This was causing the app to break.

One fix is to only load the "primary" (here, first) AA provided by a
module fixture, and rely on some combination of the projection back
through deck config to make the other AAs load and on the use of those
AAs in subsequent commands if threy matter.

A better fix is to just have the engine list the possible cutout fixture
configurations that will work for the protocol. But this is good for now.