feat(api,app): finish up new stacker API (#17620)
This PR removes static mode and updates the protocol api to use the new
way of using the stacker.

The _rest_ of the PR fixes things we missed in previous PRs, such as
- We actually have to build new `LabwareCore` entries for the labware
returned by `retrieve` in the protocol core
- When we load multiple labware in `retrieve`, we need to be able to
tell the rest of the engine about the rest of the labware we're loading,
since if we're loading labware on adapter all at once, the internals
need to know the identity and details of the adapter to figure out what
to do with the labware
- While we're not adding handling for command text for these commands,
we do need to add something so the app doesn't crash. There's initial
support for reading the data from the command results, but mostly I made
the LPC internals ignore labware that we don't know the definition for
instead of hitting the error boundary. So now labware that starts its
life in the stacker hopper just won't show up in the app at all.
- The app couldn't properly handle the new combo fixtures, because it
was eagerly loading all the addressable areas provided by a default
module feature when it saw a `loadModule`, which means there would be a
fundamental conflict if you first put a stacker in D3 (providing
`flexStackerModuleV1D4` and `D3` from the `flexStackerModuleV1` cutout
fixture, even if you weren't using `D3`) and then used a waste chute,
which would conflict with D3.

## to come out of dev
- [x] fix the tests that are failing
- [x] add new tests

## review requirements
- what do you think of the new labware core stuff
- what do you think of the new way we load labware 

## testing
I ran this protocol:
```
from opentrons.protocol_api import ProtocolContext, OFF_DECK

metadata = {
    "protocolName": "test new stacker api",
}

requirements = {"robotType": "Flex", "apiLevel": "2.23"}


def run(protocol: ProtocolContext) -> None:
    instrument = protocol.load_instrument("flex_8channel_1000", "left")

    stacker_1 = protocol.load_module("flexStackerModuleV1", "D4")
    stacker_2 = protocol.load_module("flexStackerModuleV1", "C4")

    stacker_1.set_stored_labware(
        "opentrons_flex_96_tiprack_200ul", lid="opentrons_flex_tiprack_lid", count=6
    )
    stacker_2.set_stored_labware(
        "opentrons_flex_96_tiprack_200ul", lid="opentrons_flex_tiprack_lid", count=6
    )

    def dump_rack_from_each():
        tiprack_1 = stacker_1.retrieve()
        protocol.move_labware(tiprack_1, "D2", use_gripper=True)
        tiprack_2 = stacker_2.retrieve()
        protocol.move_labware(tiprack_2, "C2", use_gripper=True)
        protocol.move_labware(tiprack_2, OFF_DECK)
        protocol.move_labware(tiprack_1, OFF_DECK)

    def swap_racks():
        tiprack_1 = stacker_1.retrieve()
        tiprack_2 = stacker_2.retrieve()
        protocol.move_labware(tiprack_1, "D3", use_gripper=True)
        protocol.move_labware(tiprack_2, stacker_1, use_gripper=True)
        protocol.move_labware(tiprack_1, stacker_2, use_gripper=True)
        stacker_1.store()
        stacker_2.store()

    dump_rack_from_each()
    swap_racks()
    dump_rack_from_each()
    swap_racks()
    dump_rack_from_each()
    swap_racks()
    dump_rack_from_each()
    swap_racks()
    dump_rack_from_each()
    swap_racks()
    dump_rack_from_each()
```

and it worked and the app didn't crash!

---------

Co-authored-by: CaseyBatten <caseyhbatten@gmail.com>