fix(protocol-designer): overhaul highlight logic for modules + labware (#17663)
This PR updates logic for which labware/module entities to highlight,
whether they be selected or hovered, and moves the logic into a utility
for code cleanliness. The logic includes checking whether a selected or
hovered labware item is a bare adapter on top of a module, in which case
we actually highlight the parent module rather than the adapter. I also
fix a bug where absorbance reader module highlights were not correctly
taken into account when opening a new or old plate reader step with
pre-selected moduleId.

Closes AUTH-1450, Closes RQA-3954, Closes RQA-3951