test(protocol-designer): Cypress-Transfer-Platereader-Helpers-Included (#17677)
This is coppied over from
https://github.com/Opentrons/opentrons/pull/17623 because my branch name
had parentheses.

# Overview

1. Add plate reader smoke test
2. Add Transfer 1:1 test
3. Add helper functions to test if we like that pattern to reduce work
used in createNewFlex.cy.ts
## Test Plan and Hands on Testing

## Changelog
Added test coverage for transfer. Might need to be refactored once we work on liquid classes. Will likely cause some issues until Jethary's PR is pulled in

## Review requests
@y3rsh 

## Risk assessment
Going to likely need to look through selectors when we get to liquid
class work because it will change fields. That being said, having
automation that can be tweaked to liquid classes is helpful.

---------

Co-authored-by: Josh McVey <josh.mcvey@opentrons.com>
Co-authored-by: Jethary <jethary@opentrons.com>