feat(step-generation): define a curryCommandCreator that discards Python (#17674)
# Overview

In step generation, compound commands are implemented by reducing over
curried `CommandCreator`s. E.g., dropping a tip in trash is defined as
something like:
```
[
  curryCommandCreator(moveToAddressableAreaForDropTip, ...),
  curryCommandCreator(dropTipInPlace, ...),
]
```
If we generated Python for each of the atomic commands individually, we
would get something like
```
pipette.move_to(...)
pipette.drop_tip(...)
```
However, in the Python API, it is more idiomatic to emit a single Python
command to drop the tip into the trash. We can do that by adding a
`CommandCreator` that emits the standalone Python command for dropping a
tip into the trash, but we need a way to *turn off* the Python for the
`moveToAddressableAreaForDropTip`/`dropTipInPlace` atomic commands. (We
can't just delete those atomic commands because we still need them for
JSON generation, plus they do useful error checking for us.)

So this PR introduces a new variant of `curryCommandCreator` that
suppresses Python from the `CommandCreator`.

So the implementation would become:
```
[
  curryWithoutPython(moveToAddressableAreaForDropTip, ...),
  curryWithoutPython(dropTipInPlace, ...),
  curryCommandCreator(pythonDropTipInTrashBin, ...),
]
```
This preserves the existing behavior when generating JSON, but emits
only a single Python command when generating Python. AUTH-1385

## Test Plan and Hands on Testing

Added unit tests.

I have also been using this in my private experimental branch for
hands-on testing.

## Risk assessment

Low. Python generation is behind a feature flag, and this is designed to
not affect JSON generation at all.