refactor(step-generation): split up movableTrashCommandsUtil() into separate functions (#17694)
# Overview

We had a single `movableTrashCommandsUtil()` function that generated
commands for air-gapping, dropping tips, dispensing, blowing out, and
moving to the trash bin. In order to select which action you wanted, you
passed in a string `type = ('airGap' | 'blowOut' | 'dispense' |
'dropTip' | 'moveToWell')`.

This was clunky and made the program flow hard to follow. It would also
make it harder to implement Python generation for these compound actions
one at a time.

So this PR splits each of those functionalities into its own separate
function.

## Test Plan and Hands on Testing

I ran the tests in `movableTrashCommandsUtil.test.ts` to confirm that
each of the functions behaves exactly as before.

## Review requests

I recommend viewing this diff with whitespace changes ignored: in
GitHub, add `?w=1` to the URL.

## Risk assessment

Low. This is a refactoring to clean up the code. There should be no
change in behavior.